Update for Meteor 0.6.5
Use api.export for exports.  (The `awwx` global variable for internal
exports is dropped entirely).

Use new versions of browser-msg and canonical-stringify which support
Meteor 0.6.5.

Add `Offline.isWebWorker` to indicate whether we're running in the
shared web worker.

Inside the shared web worker environment, use `WebWorker` as the
global for the web worker API instead of `Agent`.

For Meteor code running in the shared web worker, update patches to
Meteor packages to 0.6.5.

Since the non-minified version of a Meteor bundle is now quite large
in 0.6.5, for worker-packages.javascript bundle without "--debug" so
that it is now minified.