Fix subscription status
Ensure Offline.supported is false if we're using browser-msg and
BrowserMsg.supported is false.

Update to fixed canonical-stringify version 1.3.

When starting the shared web worker, use the worker-boot.javascript
file hash as the worker name, as Chrome wasn't starting a new web
worker when the URL only differed in the query hash parameter.  (Still
TODO is reloading if worker-packages.javascript changes).